pyproject.toml
[tool.poetry]
name = "discharger"
version = "0.1.0"
description = ""
authors = ["gokullyric <110442611+gokullyric@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
minimalmodbus = "^2.1.1"
pyserial = "^3.5"
flask = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
